+++ src/modules/edc-demo/components/asset-editor-dialog/model/http-params-form-auth-header-type.ts
@@ -1,0 +1,1 @@
+export type HttpParamsFormAuthHeaderType = 'None' | 'Value' | 'Vault-Secret';