+++ src/modules/edc-demo/components/contract-agreement-transfer-dialog/contract-agreement-transfer-dialog-form-model.ts
@@ -1,5 +1,7 @@
-import {FormGroup, ɵFormGroupValue} from '@angular/forms';
-import {ContractAgreementDatasinkFormModel} from './model/contract-agreement-datasink-form-model';
+import {FormArray, FormControl, FormGroup, ɵFormGroupValue} from '@angular/forms';
+import {DataAddressType} from '../data-address-type-select/data-address-type';
+import {HttpDatasinkAuthHeaderType} from "./model/http-datasink-auth-header-type";
+import {HttpDatasinkHeaderFormModel} from "./model/http-datasink-header-form-model";
 
 /**
  * Form Value Type
@@ -11,5 +13,23 @@ export type ContractAgreementTransferDialogFormValue =
  * Form Group Template Type
  */
 export interface ContractAgreementTransferDialogFormModel {
-  datasink: FormGroup<ContractAgreementDatasinkFormModel>;
+
+  dataAddressType: FormControl<DataAddressType>;
+
+  // Custom Datasink JSON
+  dataDestination: FormControl<string>;
+
+  // Http Datasink
+  httpUrl: FormControl<string>;
+  httpMethod: FormControl<string>;
+
+  httpRequestBodyEnabled: FormControl<boolean>;
+  httpRequestBodyValue: FormControl<string>;
+  httpContentType: FormControl<string>;
+
+  httpAuthHeaderType: FormControl<HttpDatasinkAuthHeaderType>;
+  httpAuthHeaderName: FormControl<string>;
+  httpAuthHeaderValue: FormControl<string>;
+  httpAuthHeaderSecretName: FormControl<string>;
+  httpHeaders: FormArray<FormGroup<HttpDatasinkHeaderFormModel>>;
 }